better logging

+++ backend/inference/remote_setup.sh
@@ -111,10 +111,11 @@ echo "Docker images:"
 sudo docker images
 
 
-
 # Run the inference server with the specific model name
 echo "Running opensora_api inference server..."
-sudo docker rm -f opensora_api || { echo "Failed to remove existing opensora_api Docker container"; exit 1; }
+if sudo docker ps -a --format '{{.Names}}' | grep -Eq "^opensora_api$"; then
+    sudo docker rm -f opensora_api || { echo "Failed to remove existing opensora_api Docker container"; exit 1; }
+fi
 sudo docker run -d --gpus all -p 5000:5000 -v /home/ubuntu/data:/data --name opensora_api opensora_api:latest || { echo "Failed to run opensora_api Docker container"; exit 1; }
 
 echo "Deployment script completed successfully."
